Created separated components for drawer & address
