+++ apps/partners-reference/src/applications/PaperApplicationDetails/DetailsMemberDrawer.tsx
@@ -1,0 +1,126 @@
+import React from "react"
+import {
+  AppearanceStyleType,
+  t,
+  GridSection,
+  ViewItem,
+  Button,
+  Drawer,
+} from "@bloom-housing/ui-components"
+import { AddressColsType, DetailsAddressColumns } from "./DetailsAddressColumns"
+import { Application, HouseholdMemberUpdate } from "@bloom-housing/core"
+
+export type MembersDrawer = HouseholdMemberUpdate | null
+
+type DetailsMemberDrawerProps = {
+  application: Application
+  membersDrawer: MembersDrawer
+  setMembersDrawer: (member: MembersDrawer) => void
+}
+
+const DetailsMemberDrawer = ({
+  application,
+  membersDrawer,
+  setMembersDrawer,
+}: DetailsMemberDrawerProps) => {
+  return (
+    <Drawer
+      open={!!membersDrawer}
+      title={t("application.household.householdMember")}
+      ariaDescription={t("application.household.householdMember")}
+      onClose={() => setMembersDrawer(null)}
+    >
+      <section className="border rounded-md p-8 bg-white mb-8">
+        <GridSection
+          title={t("application.details.householdMemberDetails")}
+          tinted={true}
+          inset={true}
+          grid={false}
+        >
+          <GridSection grid columns={4}>
+            <ViewItem
+              label={t("application.name.firstName")}
+              children={membersDrawer?.firstName || t("t.n/a")}
+            />
+
+            <ViewItem
+              label={t("application.name.middleName")}
+              children={membersDrawer?.middleName || t("t.n/a")}
+            />
+
+            <ViewItem
+              label={t("application.name.lastName")}
+              children={membersDrawer?.lastName || t("t.n/a")}
+            />
+
+            <ViewItem
+              label={t("application.household.member.dateOfBirth")}
+              children={
+                membersDrawer?.birthMonth && membersDrawer?.birthDay && membersDrawer?.birthYear
+                  ? `${membersDrawer?.birthMonth}/${membersDrawer?.birthDay}/${membersDrawer?.birthYear}`
+                  : t("t.n/a")
+              }
+            />
+
+            <ViewItem
+              label={t("application.add.sameAddressAsPrimary")}
+              children={
+                membersDrawer?.sameAddress === null
+                  ? t("t.n/a")
+                  : membersDrawer?.sameAddress
+                  ? t("t.yes")
+                  : t("t.no")
+              }
+            />
+
+            <ViewItem
+              label={t("application.add.workInRegion")}
+              children={
+                membersDrawer?.workInRegion === null
+                  ? t("t.n/a")
+                  : membersDrawer?.workInRegion
+                  ? t("t.yes")
+                  : t("t.no")
+              }
+            />
+
+            <ViewItem
+              label={t("t.relationship")}
+              children={
+                membersDrawer?.relationship
+                  ? t(`application.form.options.relationship.${membersDrawer?.relationship}`)
+                  : t("t.n/a")
+              }
+            />
+          </GridSection>
+
+          <GridSection grid={false} columns={2}>
+            <GridSection subtitle={t("application.details.residenceAddress")} columns={3}>
+              <DetailsAddressColumns
+                type={AddressColsType.memberResidence}
+                application={application}
+                householdMember={membersDrawer}
+              />
+            </GridSection>
+
+            {membersDrawer?.workInRegion && (
+              <GridSection subtitle={t("application.contact.workAddress")} columns={3}>
+                <DetailsAddressColumns
+                  type={AddressColsType.memberWork}
+                  application={application}
+                  householdMember={membersDrawer}
+                />
+              </GridSection>
+            )}
+          </GridSection>
+        </GridSection>
+      </section>
+
+      <Button styleType={AppearanceStyleType.primary} onClick={() => setMembersDrawer(null)}>
+        {t("t.done")}
+      </Button>
+    </Drawer>
+  )
+}
+
+export { DetailsMemberDrawer as default, DetailsMemberDrawer }
